requirements.txt
streamlit
Pillow
pandas
sqlite3
plotly-express
docx2txt
PyPDF2
pdfplumber
xlwings
requests
fitz  # PyMuPDF
pptx
google-generativeai